+++ packages/adhere-ui-searchtable/src/searchform.tsx
@@ -27,7 +27,8 @@ class SearchForm extends React.Component<ISearchFormProps, any> {
         className={classNames(selectorPrefix, ...(className || '').split(/\s+/))}
         style={{ ...(style || {}) }}
       >
-        <tbody>{children}</tbody>
+        {/*<tbody>{children}</tbody>*/}
+        {children}
       </table>
     );
   }