+++ packages/adhere-website/src/components/ui/messagedialog/index.jsx
@@ -489,7 +489,7 @@ export default () => {
         config: {
           title: '提示',
         },
-        defaultCloneBtn: false,
+        defaultCloseBtn: false,
         children: (
           <Result
             title="Your operation has been executed"
@@ -522,7 +522,7 @@ export default () => {
                 config: {
                   title: '提示',
                 },
-                defaultCloneBtn: false,
+                defaultCloseBtn: false,
                 children: (
                   <Result
                     title="Your operation has been executed"
@@ -1032,7 +1032,7 @@ export default () => {
                     required: '',
                   },
                   {
-                    name: 'defaultCloneBtn',
+                    name: 'defaultCloseBtn',
                     desc: '是否缺省有关闭按钮',
                     type: 'boolean',
                     defaultVal: 'true',