+++ packages/adhere/changelog/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.0.17
+
+***
+
+2022-6-2
+
+* **adhere-ui-messagedialog**
+  - Modal参数名称修改(cloneBtn修改为closeBtn)
+* **adhere-ui-searchtable**
+  - SearchForm的children位置修改
+* **adhere-util**
+  - tree的set...修改
+
 # 1.0.16
 
 ***